+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -154,5 +154,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-29T13:06:54.086Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-29T14:44:54.875Z")
     r"""The date and time the resource was last updated"""