+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6ASAWPPJAQGCBP1H4A2GMJV"
+        "01K6AYYBBQS36HSPZ3HWAH3SGD"
     )
     r"""The id of the resource"""