+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6ASAWK8RAV80C5PT0WV1HS6"
+        "01K6AYYB89J93VJTE3355JJ7VQ"
     )
 
     requires_approval: Optional[bool] = False