+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-29T13:06:54.086Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-29T14:44:54.875Z")
     r"""The date and time the resource was last updated"""